Start server only after the first emit
Problem:
The server is started before webpack emits the initial changes and
built files are written to disk, causing the installed extension,
which is trying to reconnect, to reload without having files to load.
This caused an error in the extension, where it needs to be removed
and installed again.

Solution:
We start the server after the first webpack emit event

Fixes #51